refact(NA): changes according PR review
